requirements.txt
ultralytics>=8.0
opencv-python>=4.5
pillow>=9.0
numpy>=1.20
scipy>=1.7
tqdm>=4.60
pyyaml>=5.4
torch>=1.9
torchvision>=0.10
